:art: 移动了部分UI文件的存储路径，更加易读
